+++ HISTORY.md
@@ -11,3 +11,7 @@
 * 2025-03-02:
 	* Updated: Dependencies
 
+* 2025-07-11:
+	* Updated: Dependencies
+	* Improved: Listing devices
+	* Improved: Minor formatting bugs